requirements.txt
requests~=2.31.0
Flask~=2.1
transliterate~=1.10.2
python-dotenv~=1.0.0